misc/HtmlTestSuite/results/test-1066.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml" dir="LTR">
<head><title>document.dir</title>
<link href="https://html.spec.whatwg.org/multipage/#dom-document-dir" rel="help"/>
<link href="https://html.spec.whatwg.org/multipage/#reflect" rel="help"/>
<link href="mailto:ms2ger@gmail.com" rel="author" title="Ms2ger"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><div id="log"/>
<script>
test(function() {
  assert_equals(document.dir, "ltr");
  assert_equals(document.documentElement.getAttribute("dir"), "LTR");
}, "Markup attribute")
test(function() {
  document.dir = "x-garbage";
  assert_equals(document.dir, "");
  assert_equals(document.documentElement.getAttribute("dir"), "x-garbage");
}, "Setting the idl attribute to a garbage value")
test(function() {
  document.dir = "";
  assert_true(document.documentElement.hasAttribute("dir"), "Attribute should still be around");
  assert_equals(document.dir, "");
  assert_equals(document.documentElement.getAttribute("dir"), "");
}, "Setting the idl attribute to the empty string")
</script>
</body></html>
